new way of handling data
only need to enter it into the customsale of all relevant planets and it should work
this needs proper readjusting of the code too (I want to see it work first, optimisation and beauty always last)

+++ source/MapPanel.h
@@ -141,8 +141,8 @@ protected:
 	std::string tooltip;
 	WrappedText hoverText;
 	// Minimum and maximum color ranges used by mapColor.
-	static double minColor;
-	static double maxColor;
+	static float minColor;
+	static float maxColor;
 	
 	
 private:
